Apply stream reference_time to individual output records
Previously, a stream's reference_time only applied to determining the
filename breaks.  Right now, output is always relative to the time of
the start of the current simulation (even it it is a restart). There is
no way to force, say, monthly output (i.e. on the 1st of every month) if
your restart does not start on the first of a month.

This commit extends reference_time to also apply to
determing when individual records should be written.  Thus, output
stream behavior will not change after a restart.

This is done by adding a new routine to mpas_timekeeping that adjusts
the prevRingTime of an alarm to be an integer multiple of the stream's
output_interval away from the stream's reference_time.  This routine is
called for each output stream created by the stream manager.