Write restart_timestamp after all output is complete
This commit moves the writing of the restart_timestamp file to after ALL
output has been written.  Previously, the restart output and
restart_timestamp file were written first, followed by other
output-direction streams.  However, I decided it makes more sense to not
consider a time step complete from a restart point of view until all of
the output is successfully written.  This is because there are some
fields unavailable on a restart, and this can results in 'missing' (i.e.
zeros) fields for some diagnostic variables if output occurs on a
restart.  Therefore it is better to 'keep' the output from the full time
step rather than have it clobbered by the restart writing.  This follows
on the previous commit that disables output from being written
on the initial time during a restart.  Thus, we do not want the
restart_timestamp file to indicate a valid restart until afer all output has
been written successfully.