Fix compilation error for SCM with prescribed aerosol plus h0 update
This is to fix compilation error for SCM when prescribed aerosol option is used (issue #1795).

The issue is introduced by PR #1723 for restructuring and reducing h0 default variables.
The calculations in mo_chm_diags.F90 for total mass mixing ratios for each class of aerosols
are only applicable when modal aerosol modules are used. The related statements are now only
enabled when -DMODAL_AERO is used in compilation.

Also included are adding CONCLD to h0 default when running V0 configuration, as well as
OMEGA500 and Lower Tropospheric Statbility (LTS, going with TH7001000 in output).

[BFB]