Change truncate to overwrite on most ocean and ice streams
The files in ACME streams.ocean have
clobber_mode="truncate"

and when I run it with truncate in stand alone, for files with more than one time like highFrequencyOutput, it will delete the whole file upon restart.  It is safer to just use
clobber_mode="overwrite"
for everything.  That does not delete the file, but just overwrites a record with the same time.  It only writes a warning to the processor zero log file, so does not create a log file for every single processor.

For simulations with yearly restarts in ACME everything is fine.  For monthly or other restarts, it will delete some analysis data.

All those lines are originally from an old commit:
c79031ad (Robert Jacob     2016-07-25 00:03:30 -0500 569)
print $stream_file '        clobber_mode="truncate"' . "\n";