Fix CFL check when adaptive timestep is off
The current code was skipping the CFL checks when the adaptive
timestepper was off (which is the only time they are really needed).
This bug appears to have been added with
9cea0f5fa7eb55c20e148f00d2fdef13ec2522bf.
The changes responsible also led to potential usage of dtSeconds before
it was initialized.

This commit fixes the bug.