Disable performance archiving if timing dir. can not be written into
Currently, if the directory where performance data and associated
provenance data are to be saved already exists or can not be created
or populated, then the job aborts. This change allows the job to proceed,
simply disabling the performance data archiving. This will allow the use of
default locations for saving the performance data that are legal for
users in the E3SM project but which will not be legal for E3SM users who
are not part of the project. These external users can define alternative
locations in their job cases.

[BFB]