Change SAVE_TIMING_DIR default at NERSC to central location
Currently the default location for SAVE_TIMING_DIR on Edison,
Cori-Haswell, and Cori-KNL is /project/projectdirs/$PROJECT .
There are a number of ACME-project allocations at NERSC, and
it is advantageous for the performance data for all of these to
be archived in a single location. Here this default is set to
/project/projectdirs/acme . If the ACME model is run by
someone not in the acme group and if this default is not
changed in env_run.xml, then performance data archiving
will be disabled.

[BFB]