Change mach_syslog for Cori to start checkpointing earlier
Currently the scripts for Cori-Haswell and Cori-KNL
that monitor model progress do not start until the number of lines
in acme.log exceeds the number of MPI processes. This design was
introduced when the process-to-core mapping was output to acme.log.
This mapping output has since been disabled for these systems
and the script often waits excessively long for jobs with large MPI
process counts. This commit changes these scripts to start after an
empirically determined number of lines, attempting to start after the
model output starts, thus after the list of MPICH environment variables
is output. As this is emprically determined, it may need to be adjusted
again in the future.

[BFB]