Change mach_syslog for Edison to start checkpointing earlier
Currently the script for Edison that monitors model progress does
not start until the number of lines in acme.log exceeds the number of
cores in the allocated nodes. This design was introduced when the
process-to-core mapping was output to acme.log. As the number of cores
can be larger (and potentially much larger) than the number of MPI
processes when using OpenMP threading, the script often waits excessively
long for jobs with large nodes counts when OpenMP threading is used.
This commit changes this script to start after the length of acme.log
exceeds the number of nodes. While not guaranteed to capture all of
the process-to-core mapping, this change does guarantee that something
is captured before the job ends. Note that this change is needed now
because of the successful cleanup of acme.log, significantly shortening
its length compared to that generated by earlier versions of the model.

[BFB]