Eliminate adding negative time to timing events
On Cori-KNL MPI_WTIME does not always return monotonically increasing
values, leading to negative time increments being added to profile
timers (for fine grain timings). This is a rare event, and the timing
library does generate warnings when it occurs. However, it still adds
in the negative values. This modification sets negative increments to
zero, decreasing the impact of this erroneous MPI_WTIME behavior.

[BFB]