Force output of checkpoint timing file after timestep 1
Performance of the first timestep typically differs from that
of the other timesteps. It is useful to capture this as
a checkpoint timing file. This also increases the odds of
capturing the cost of the initialization even when the
initialization is very expensive and the job exceeds the
wallclock limit and is killed before final timing data
is output.

[BFB]