Reverted rain fallspd calc to pre-sed version + commented on sed evap
Chris' rain fallspeed calculation performed explicitly what
size_dist_param_basic() was meant to do. I think this was for efficiency,
but in that case it's unclear why this approach wasn't used for snow sed
as well, since snow sed is handled identically. I also think Chris'
implementation had a slight bug where "**br" was moved into the lam
calculation rather than the fallspd calc, but it wasn't applied to the
same terms as in the original formulation.

Changes to be committed:
	modified:   micro_mg2_sedimentation.F90