Change paths to case2 rundir and exeroot
Previously, the code assumed that rundir and exeroot were subdirectories
of CIME_OUTPUT_ROOT. This is the case for CESM machines, but not for all
ACME machines. The new version doesn't make this assumption.

Also fixes a minor bug with case_fake: in cloning, set the clone's
cime_output_root and rundir rather than the original case's.