Add unit tests of compare_two's handling of pass/fail in comparison
Previously, I had figured that it was sufficient to ensure that
_component_compare_test was actually being called, figuring that the
tests of _component_compare_test belong elsewhere. But, since this is
such a critical aspect of the system_test_compare_two infrastructure,
I'm adding some tests covering _component_compare_test here.

Note that I have now removed the checks of the "Call" objects to
_component_compare_test. Now that we're calling and testing the actual
_component_compare_test, it is no longer possible or necessary to
construct and check a "Call" object recording this call.