simple_condensation_model.F90: add more output; clean up
A number of quantities calculated in the simple condensation model are added
to model output.

During this process, a small clean-up and re-organization was done to
facilitate the diagnosis and make the code structure (hopefully) easier
to understand.

More comments are added. The names of some tmp arrays are changed,
again to improve readability.

Test runs were conducted using 3 configurations:
 - RKZ_A1_B0_C2_ql17
 - RKZ_A1_B1_C0_lmt4
 - RKZ_A1_B1_C1_ql7
using 1 ensemble member and 2s/1s time step. The results are unchanged
judging from the RMSEs.