Changes code to read "date" field instead of "month"
"date" field (with "date" dimension) has been added to the input data
file to be consistent with the rest of the data files. Code has been
also changed to accomodate this change in the data file.

Also modified the temporary model ouput for sw extinction from 1/m to
1/km so that it is easy to compare to the same field in the data
file. Now extinction for both LW and SW has same units of 1/km.

Note that this change doesn't change answers except the temporary
fields (outputted read in and interpolated LW and SW extinction)

[BFB] - Bit-For-Bit