Update to CMIP6-1850: Strat Aerosol now uses DATE.
The previous version of the stratospheric aerosol file contained its
time information in 'month'.  However, this meant overwriting the
original file data (which used months since 1850).  So, to make it
cleaner and more consistent, the variable 'date' was added to the data
file instead of modifying 'month'.  However, this required
modification of the code to read the new variable. Those changes were
in the previous commit to this branch, but those changes assumed that
'date' would have its own dimension, so this commit changes the code
so 'date' uses the 'month' dimension.  This will avoid problems if the
data file is every modified.

The CMIP6_1850 compset has also been updated to use the new file.

[Should be BFB]