simplify implementation of is_save_timing_dir_project
The implementation of is_save_timing_dir_project was simplified based
on reviewer suggestions. This also introduced support for standard
regular expressions, and ANY was replaced with '.*' as the wildcard
for SAVE_TIMING_DIR_PROJECTS in config_machines.xml .

The descriptions of SAVE_TIMING_DIR and SAVE_TIMING_DIR_PROJECTS
were also tweaked.

[BFB]