Add option to pass all cloud as stratiform to COSP
Add option to pass all cloud and precipitation properties as stratiform
to COSP. This is needed when the deep convective scheme is turned off,
because the convective fields are only defined in within the deep
convective scheme. So when we want to run CLUBB-SILHS with
deep_scheme='off', then runs with COSP will fail when it tries to look
for these. The new behavior is to look for the presence of these fields,
and if they do not exist then all cloud properties will be passed as
stratiform. A namelist variable is added (cosp_separate_convection) to
control this in runs that do have deep_scheme on for debugging purposes.