Remove unused doctests
While profiling, I inadvertently discovered that the import of doctest
is actually somewhat expensive computationally. We had copy/pasted a
pattern of supporting a --test option for all our exes but --test was
rarely doing anything. We currently have a good platform for running
doctests in scripts_regression_tests, so there's no need to try to
duplicate this in our exes.