Add aprun for the jobs that are compiled in mpi-serial on Titan
Because the CPU processors in Titan's login, sevice and compute nodes
are different, the "aprun" command is always required even if the
model is compiled in mpi-serial. This PR is to add the "aprun" command
before "acme.exe" in the batch script on Titan when the model is compiled
in mpi-serial. It will fix the issue #2022

[BFB] - Bit-For-Bit