Fixes NAG identified errors in atm, homme and clm models
Most of changes are related to line length longer than 132 characters,
which NAG doesn't allow. Other changes are related where the whole F90
files were enclosed in #if defines. NAG complains about these files
as it sees no code in those files when those #if defines are not
true. Previously Gautam identified that these kind of files may
force recompiliation of the some codes even if no change is made to
any of the F90 files.

[BFB] - Bit-For-Bit