NERSC is changing the way to submit jobs. The old way still works now, but this change will use the new way. Added a new batch system "nersc_slurm" to contain the one change needed. They want to use -q regular/debug and eliminate -p (for partition) and --qos.
Minor module version updates to system defaults for cori-knl and cori-haswell.
Make the default compiler on cori-haswell be Intel version 18 (as on cori-knl).
No changes to edison modules, except for module version of the intel18 compiler option (now if --compiler=intel18 is used, should get the same version/flags as on cori)

Removed custom setting of PIO_STRIDE=64 for cori-knl. Now the stride should be set to the number of MPI's per node as with most other machines.

The default GCC version at NERSC is now 7.1, but there were several LAND build issues, so the version was left at previous version of 6.3 for all 3 machines.

I also adding a module command to remove the Cray Programming environment if it was loaded for cori-knl, which would Fix github 1840

Added module remove statements for cray-libsci before loading specific versions which was required for cori-haswell & cori-knl when using GNU compiler (Fix githib 2039)