Fix incorrect sign in boundary relaxation terms for rho, rt, and ru
The sign of the contribution to the tendendies for rho, rt, and ru due to
relaxation towards driving values along the domain boundaries was incorrect due
to the reversal of the terms in difference between the model values and
the driving data values. This commit corrects the sign of these terms by
swapping the order of the terms in the difference computation.