Remove lbc_rho_edge from LBC forcing files, and compute this field internally
The 'lbc_rho_edge' field is not needed on the outermost edges of a limited-area
domain, so we can compute this internally by averaging lbc_rho_zz to edges.
The results when averaging lbc_rho_zz to edges in the mpas_atm_boundaries
module will differ from those when reading lbc_rho_edge from the LBC forcing
files, since the value of lbc_rho_edge in those forcing files is not averaged
from the cells on either side of an edge, but rather is interpolated directly
to edge locations from the cell-centered rho_zz field in the driving model mesh.