Switch from 'lbc_theta_m' to 'lbc_theta' in LBC forcing files
We now read 'lbc_theta' from the LBC forcing files and use this internally to
compute lbc_rtheta_m using lbc_rho_zz and lbc_scalars(qv). The lbc_theta_m field
is apparently not used anywhere in the solver, so there is no need to retain
this field in addition to lbc_rtheta_m.