Add eigencalving option
This commit implements the “eigencalving” calving law described in:
   Levermann, A., T. Albrecht, R. Winkelmann, M. A. Martin, M. Haseloff,and
   I. Joughin (2012), Kinematic first-order calving law implies potential for
   abrupt ice-shelf retreat, Cryosph., 6(2), 273–286, doi:10.5194/tc-6-273-2012.
In this law, calving rate is proportional to the product of the two principal
strain rates if both are extensional.

The calculation of the calving rate is implemented in a subroutine called “eigencalving”.
It makes use of two helper routines (that could potentially be used in other ways):
1. calculate_calving_front_mask - makes a mask of floating ice that is the last dynamic cell
before the calving front
2. distribute_calving_flux - takes a specified volume calving rate and
distributes it to the cell in question and its neighbors.  This routine has one limitation in
that it only considers immediate neighbors, and therefore can fail to distribute all required
calving volume. The fix would be to make the routine recursive, but in practice this
limitation may not be very significant (no conservation laws are broken).  For now, to allow
assessment of if and where this occurs, a warning is written on each processor where
this is a problem, and the diagnostic output field “uncalvedVolume” indicates where and how
much volume was failed to get calved.  A global stats member could be created to
track this more conveniently.