Fix bug in extrapolation of floating b.c.
The Albany velocity boundary condition is extended at floating margins
to use the condition of the thinnest neighbor with ice.
However, this was supposed to use the thinnest *dynamic* neighbor with
ice, but the logic inadvertently was using any neighbor with ice.  This
resulted in the extended locations on the Albany mesh often being thinner than
they should be because non-dynamic neighbors tend to satistfy the
criterion of thinnest neighbor.

This commit fixes this by only considering dynamic neighbors as the
source of state information to be used in the extension.