Exclude thin floating margin cells from being dynamic in mask
This commit changes the definition of what a dynamic cell is at the
margin of ice shelves.  Previously, the dynamic thickness limit was used
on both grounded and floating ice.  This commit marks cells along the
margin of floating ice as non-dynamic unless they are at least 95% of
the thickness of their ice shelf-interior neighbors.  This results in a
"poor man's calving front advection scheme".  The ice shelf front cannot
advance until the cell at the margin is "full" - as thick as the ice
behind it.

This was added to allow the eigencalving scheme to partially deplete
cells along the ice shelf front without doing weird things to the
geometry there.  But it seems like a sensible and preferable choice
under all conditions.