Always call rbf init routines
These routines were lazily skipped when the FO velocity solver was being
used because the MISMIP3d meshes were not set up correctly and were
causing an error in the rbf init routines.

The fix should be to fix the MISMIP3d meshes, not to hobble the code.
The rbf routines are needed to calculate strain rates on cell centers
needed for the eigencalving scheme (which is used with FO velo solver).

So this commit makes the rbf init routines always called.