Leverage env command options to assist parsing
Newline was not a strong-enough split character for parsing
output of 'env' since newlines can occur in environment variables,
especially functions. Incorrect handling was causing all machines using
the 'soft' environment manager not to work.

This change causes a null character to be placed between env variables,
which should be a much more reliable way to split/parse this output.