Add enhancement factor field and prepare to pass it to Albany
This commit creates the field and sets it up to be passed to Albany, but
it does not yet populate the field or apply it to the other parts of the
code (SIA velocity, thermal solver, hydro model).  Also, the passing of
the new field to Albany is currently commented out until a known version
of Albany to support it is available.