Removes a land F90 file that is autogenerated
dynVarTimeUninterpMod.F90 is auto generated from dynVarTimeUninterpMod.F90.in
and should not be tracked. Prior to the merge of PR #2024, tracking
dynVarTimeUninterpMod.F90 did not cause any build failures
because the auto-generated dynVarTimeUninterpMod.F90 was exactly similar
as the tracked dynVarTimeUninterpMod.F90. PR-2024 made changes to
dynVarTimeUninterpMod.F90.in, thus dynVarTimeUninterpMod.F90
should not be tracked.

Fixes #2102

[BFB]