Initialize enhancement factor and apply it to rest of code
This commit continues from the previous commit and initializes the
enhancement factor using one of two methods (scalar value or data from
file) and then applies to other parts of the code where flowParamA is
used: SIA velocity, thermal, subglacial hydro.
This commit removes the application of the enhancement factor in the
calculation of the flow rate parameter, instead keeping E separate from
A and explicitly including both factors wherever A is used.