Refactor handling of small jobs
We already allowed a job to declare that it only needed 1 task. We
should also allow it to declare a smaller walltime. When this two
abilities are combined, it should allow for the selection of the
appropriate "small job" queue if the machine has been configured
correctly.

Also, get rid of unused pedocumentation.