For Travis-CI build: kick off build on master whenever it's pushed
The Travis-CI build was already happening on master, but it was also
happening on ALL branches pushed to ESMCI, which was
overkill. @jedwards4b changed the configuration so that it wouldn't run
on pushed branches in general. This change (if I did it right) restores
its running on master, which seems like a good thing to have.