Minor changes to patch updater
The ratio of change in area for a patch is now computed as difference in
patch weight w.r.t to a grid cell divided by old weight of patch w.r.t to
grid cell. Previously this ratio was computed using weights w.r.t to column.

Changes in seed and flux are computed using chnages in weights w.r.t. to
column.