Bug fixes for dynamic root model
This commit updates the dynamic root model to fix the
spinup delay in the tropics. The updates included a revised
water profile that matches btran, an exponential N profile when
vertical profiles are not on, and allows changes to the root profile
during the growing season rather than just the onset period for
stress and seasonal deciduous vegetation. The dynamic roots also
maintain the coarse root distribution.

[non-BFB] when use_dynroot = .true.