When deriving QV from RH, always assume RH is w.r.t. liquid water
Under the assumption that real-data intermediate files were produced by
the WPS's ungrib program, we should always assume that the RH field is
with respect to liquid water, since this is what WRF's real.exe program
assumes and what the WPS ungrib.exe produces.

Thanks to The Weather Company for finding this issue.