Fix AMs for floating BMB to only calculate it for floating ice
If the BMB field exists in places where the ice is grounded, the current
analysis member calculations related to floatingBasalMassBal will
include those values in the calculation.  This commit restricts the
calculation of related stats to just floating ice.

In the long term, it would be more accurate to add a new model field
(e.g. floatingBasalMassBalApplied) that stores only what value was
actually applied - this would also catch against the situation where the
prescribed BMB was greater than the amount of ice in a cell.  Similar
terms for SMB should be added.