Updates to eigencalving: thickness-threshold, mask definitions, stranded ice
This commit includes a number of minor but significant adjustments to
the eigencalving functionality.
1. Change thickness-threshold to only apply to cells on the calving front.  This required
updating the geometry and mask before applying the thickness threshold. That in turn
required rearranging how the data-calving option is implemented in the calving module,
because eigencalving needs to make multiple 'passes' of calving,
applying it to the thickness and mask each time.
2. Mask definitions in the mask and calving modules have been updated.
Specifically, if a dynamic floating margin is adjacent to non-ice land,
then it should not be treated as a calving margin.  Before this fix,
holes were developing in shelves near the grounding line in some place
of relatively stagnant flow.  Once established, these holes would
eventually grow to consume the entire shelf from the inside out.
3. Eigencalving now includes an operation to remove abandoned floating
ice that can get cut off from the shelf (i.e. icebergs).  The cleanup
operation is not completely effective at removing icebergs immediately,
but in practice find most of them, and what it does not find immediately
is often garbage collected later as they thin from basal melting.

All of these changes were required to get AIS 20km spinups with
eigencalving to behave stably and reasonably.