Fix environment var settings
Recent cime updates have made empty-string env settings to mean
that the env var should be unset.

This commit changes all occurences of e3sm setting vars to empty-string
to set it to a single space instead, since in none of the cases did we
intend to unset the var.