Fix for MPI_Alltoallv with unallocated bufs
In rearrange_ the buffers ISendBuf, IRecvBuf, RSendBuf, and
RRecvBuf are allocated only when they are needed to send
or receive integer or real data. When usealltoall is .true.
and useswapm is .false., then the MPI collective MPI_Alltoallv
is used to implement the underlying communication pattern.
For some MPI library implementations, this call will fail if
either the send or receive buffer is not allocated, even if
the corresponding buffer will never be used. This modification
forces the buffers to always be allocated, with a size of at
least one integer or real.

[BFB]