COSP-MODIS bugfix with APPLY_POST_DECK_BUGFIXES directive
Fix an error in COSP 1.4 MODIS simulator

This update fixes an error in the calculation of NIR reflectance that affects cloud fields derived from MODIS simulator.
As a tentative arrangement, the fixes are only effective when the CPP directive -DAPPLY_POST_DECK_BUGFIXES is applied to
CAM_CONFIG_OPTS, which can be specified in compsets of interest in components/cam/cie_config/config_components.xml.

The impact is generally small.

Modified:   components/cam/src/physics/cosp/MODIS_simulator/modis_simulator.F90

BFB for all that are not output from MODIS simulator when the CPP directive is defined.