Improve runtime of PET_Ln9.ne30_oECv3_ICG.A_WCYCL1850S integration test
This test is typically, by far, the slowest of our integration tests
to complete. Since we are only doing a Ln9 run, it makes no sense to
use a large peylayout. A much better approach is to use the PS test option
to force a smaller layout so we can get in the queue faster.

This PR also defined a small grid layout for sandiatoss3.

This PR allows both phases of the above test to fit into the sandiatoss3
"short" queue.

[BFB]