Remove exner_base, pressure_base from IC file, compute these during start up
The exner_base and pressure_base fields can be computed from other fields
during a model cold-start, so we can eliminate these from the IC files.

Note that this commit does introduce changes to results due to round-off
differences in the re-computation of exner_base and pressure_base in
the atm_init_coupled_diagnostics routine.