Fix incorrect size of 'hlanduse' array in GWDO field computation
The 'hlanduse' field used in the computation of some GWDO sub-grid-scale
orography statistics was allocated with size nCells. However, this field
is accessed as hlanduse(cellsOnCell(i,iCell)), with iCell=1,nCells,
leading to out-of-bounds memory references for the outer layer of halo
cells. The hlanduse field is now allocated with a "garbage" cell, i.e.,
it is allocated with size nCells+1.