Implement multiple identical topounits (max_topounits > 1)
Replaces the use of gridcell-level landunit_indices() with
topounit-level landunit_indices(). This commit is BFB with
max_topounits = 1, but has two non-BFB problems when max_topounits > 1.
One problem is a minor issue with diagnostic output PCT_LANDUNIT. The
Other problem is more complex, involving dyn_pft logic, and will need to be
fixed on the next commit.

[BFB]

L2-178