For dlnd, read data model domain fractions from first stream file
For dlnd operating in GLC_CPLHIST mode, it's important for dlnd to have
the correct domain fractions to be able to reproduce a run done with an
active land model. However, the logic invoked for cplhist mode, where
domain information is read from the first stream file, hasn't been
reading these fractions. This commit adds the ability for
shr_strdata_init to read a data model's domain fraction from the first
stream file, and exercises this ability for dlnd.

The way this is implemented is not ideal. See comments in
ESMCI/cime#2515 for thoughts on a more general and robust
solution. However, @mvertens and I felt that we should go with this less
invasive solution for now.