Add ability to skip checks of some locked env files.
We need to lock env_batch in case.setup so that we know if we need to
regenerate submit scripts in case.submit, but we do not want case.build
barfing on env_batch.xml lock files. Thus, the concept of skipping certain
files in check_locked_files had to be added.