tweaks to homme.py to propery run HOMME tests
This patch updates homme.py so that when the HOMME test in run in CIME
without -c or -g, it creates its own baselines (make baseline), and then runs the
internal HOMME tests (make check) that include ERS, PEM and PET type tests.

Currently, the HOMME tests requires baselins before they will run, so this is the
best way to ensure the HOMME test on systems without baseslines is actually
testing something.