save e3sm.log and cpl.log files in performance archive
The recent change in CIME not to copy the component and e3sm
log files back to $CASE/log had the side effect of also
preventing the CPL and E3SM log files from being saved
in the performance archive. This modification corrects this,
copying the log files instead from the run directory.

[BFB]