Remove initialization of tbot to posinf
According to Brian Eaton, this code was added (a long time ago)
to replace code that was supposed to be trying to initialize
TBOT from the initial data file, but was actually broken and doing
nothing. This replacement code, setting tbot to posinf, was assumed
to do no harm. It was recently noticed that setting INFO_DBUG > 1
causes seq_diag_avect_mct to attempt to sum TBOT while it is still
a vector of INFs, which can cause the code to abort in certain
circumstances. Commenting out these lines eliminates this issue.
Note that TBOT is already initialized, to zero, along with all
the other surface state variables in camsrfexch.F90, so summation
of TBOT when INFO_DBUG > 1 is now a vector of zeroes.

At some point TBOT must be reset to something other than posinf,
as the posinf values do not cause problems in the simulation.
In particular, with this change (leaving TBOT with the
initialization to zero), nothing changes in the simulation.

[BFB]