Initial stub commit for adding an iac component and GCAM
This is the initial commit for new code that stubs in an integrated
assessment component (iac).  Changes were made to the primary coupler
module cesm_comp_mod.F90 to include standard send/run/recv calls for the
iac component and for coupling with lnd (input and output) and atm
(output).  The new functions called by cesm_run() and cesm_init() were
created and stubbed, and some (although probably not all) of the coupling
functions that need to understand components were modified to include
similar iac interactions.

I also moved the GCAM code from the previous commit to a (hopefully) more
standard directory structure, so there are a bunch of "rename" elements to
this commit.

GCAM itself has not been modified and cannot yet be called or included. The
goal of this commit was to outline the structure of future changes and get
most of the way to being able to build and run existing tests and cases
with the iac stubs in place.