Remove MG2 nsubr conservation limiter.
Removes a limiter that (a) was incorrectly coded, and (b) would have
been redundant if correctly coded.

The original code contained three lines that:

1. Calculated the fraction of rain mass being evaporated in the current
   time step.
2. Limited that fraction to no more than 1.
3. Evaporated the same fraction of the rain number.

Unfortunately, these lines did not use the precipitation fraction to
convert between in-precip and grid-mean quantities. In any case, the
limiter is redundant with both the mass conservation limiter above it,
and the number conservation limiter below it. When the limiter is
removed and the code is simplified, the number evaporation rate is
simply set to remove the same fraction of rain number as the mass
evaporation removes from the rain mass.

This change is potentially climate changing, since it systematically
increases rain drop size.

Fixes #2415.

[CC]