Move reading of radiation namelist
Move reading of radiation namelist to individual radiation packages.
This places the responsibility of reading the namelist options on the
specific package in use. The purpose of this is to prepare for importing
the RRTMGP radiation package, which will use some namelist options that
are irrelevant for the other radiation packages. Other namelist options
are already obsolete for the current implementation (iradae), so it
makes little sense to continue building in backward compatibility with
ancient radiation packages into runtime_opts. This commit will cause a
NLFAIL since the namelist categories have changed (the radiation
namelist is now moved from cam_inparm to radiation_nl).