Move is_cmip6_volc flag to module data
Move the is_cmip6_volc flag to module data to prevent passing it through
multiple levels of subroutines just to finally use it in the aerosol
optics routine. The purpose of this is just to simplify the calls to the
radiation interface, and keep the connection between the prescribed
volcanic aerosol mode specific to the aerosol optics interface (the only
place it is actually used).