Provides sampling for LIGHT particles
Note, this is NOT OPTIMIZED FOR IO and these fields are always
communicated in the halo.

In order for this to be optimized these fields only need transfered for
writing out the fields, not for communication, which means they are
effetively nonhalo variables. Nonhalo variables were previously
deactivated for simplicity.

Temperature and salinity sampling default to true in registry until
the E3SM registry is updated.

Add LIGHT sensors
-----------------

 * Change to particleTemperature and particleSalinity to avoid confusion
   with MPAS-O output
 * Add particle* for Temperature, Salinity, DIC, and ALK, etc
 * Update units for particle sensor variables
 * Updates 20km ziso default case to test particles
 * Tests new functionality for particleTemperature and particleSalinity
 * Refractors identification of LIGHT vertical index
 * Improves LIGHT scalar interpolation

BFB against ocean/develop (old commit 2612aed3a81cb9ad7e9aade5b3d00b63c986a14a)

BFB against baseline and ocean/develop (except for zLevelParticle
because particle position clipped to be at least dz/2 away from
boundary) with old commit 7643db31275e94f6beef873d986e677c05be38f2

Uses horizontal interpolation for sampling of scalars for particles
-------------------------------------------------------------------

Previously, nearest vertex selection approach failed because of
non-homogeneity in the mesh.  The existing approach takes a brute-force
approach and finds the mesh when neededing to do horizontal
interpolation.

Note, this is a far from optimal solution and it does not consider
corner cases where the vertex is on the boundary.  But, need to ensure
that we have adequate cells on the halo to perform horizontal
interpolation to the particle position.

Note: The following floating-point exceptions are signalling:
IEEE_DENORMAL

Ensures trajectory invarient to scalar interp via temporary
variable for zLevelParticle.

Previously, for horizontal advection near a boundary,
particles would essentially use a terrain following
coordinate for horizontal spatial interpolation, but
a local variable wasn't used for this process
and it artificially introduced spurious vertical motion
into the particle trajectories.

Adds testing for particle sensors to SOMA (T,S)
-----------------------------------------------

The SOMA initialization is from analytical functions varying
in the vertical for temperature and salinity and this test
uses particles to sample the temperature and salinity for
the initialized fields from SOMA and compares these results against
the expected initialization functions in an analysis step.

~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~

Passes BFB against SOMA restart, regression suite (except for T,S and
zLevelParticle as expected), and paraview-based visual comparison
of interpolated values