Add LC machines dependency option and debug queue.
The LC machines' version of the "slurm" batch system uses the same
commands/options to queue up a dependent job as other slurm systems, but
this information was out of date or missing, so the relevant parts of
config_batch.xml had to be copied to the lc_slurm section.

This update also adds the "pdebug" queue, which is used for small
debugging and test runs.