Complete overload of baseline comparison in MVK test
Because MVK  is doing a statistical ensemble analysis of the simulation,
it only makes sense to run the comparison once all model runs are
finished. However, since CIME runs the comparison at the end of each job
submission, we need to override (invalidate) the `-c/--compare` command
line option to `create_test` in order to get the test to run as we'd
like. Every time MVK is run, whether `-c` is given or not, the
`COMPARE_BASELINE` is set to false during the test initialization unless
`resubmit == 0` and `GENERATE_BASELINE` is false, and then it is set to true.

This means that every time this test is run without generating a
baseline, the comparison will be run, which makes sense -- there is no
reason to run the comparison when generating the baseline, and there is
no reason to run this test without either generating the baseline or
running the comparison.

Note: it isn't sufficient to simply only run the comparison if `resubmit
== 0` because the resubmit value is decremented before
`_compare_baseline` is called causing the comparison to be launched one
submission too early. Thus, we need to not call `_compare_baseline`
until we're on the very last job submission, which is achieved by
setting the `COMPARE_BASELINE` flag at initialization (done at the
beginning of each job).

The comparison will generate a JSON config file for EVE, launch EVE,
analyze both the baseline and the test simulation, and provide a
pass-fail result (by parsing the JSON data provided in EVE's generated
output website).