Generate a meaningful error msg when endrun called with no parameters
There are two general ways that endrun is called in CAM, one with a
string argument that describes the error causing the abort and one
without any parameters, but with an explicit write describing
the error that precedes the call to endrun. Examples of the two are as follows.

     if (pcols < maxblksiz) then
        write(iulog,*) 'pcols = ',pcols, ' maxblksiz=',maxblksiz
        call endrun ('PHYS_GRID_INIT error: phys_loadbalance -1 specified but PCOLS < MAXBLKSIZ')
     endif

and

     else
        if (masterproc) then
           write(iulog,*) "PHYS_GRID_INIT error: opt", opt, "specified, ", &
              "but could not divide processes into pairs."
        endif
        call endrun()
     endif

endrun is just a rename of shr_sys_abort, which itself is just a rename
of shr_abort_abort. When endrun (or shr_abort_abort) is called without a
parameter it outputs the following string:

  ERROR: Unknown error submitted to shr_abort_abort.

This occurs for each process calling endrun.

When endrun is called in the second fashion, a meaningful message is written
to atm.log by masterproc, but then shr_abort_abort prints out npes-1 copies
of the "Unknown error" message to e3sm.log and one copy to atm.log. Unless
the user is familiar with this behavior, they may not think to look for
a "real" error message in one of the component logs.

It would be nice if a call to endrun without a parameter would not produce
any output, since there is already an explicit error being output. This would
require a change to shr_abort_abort and would affect the whole model, so
will require further investigation. This PR instead modifies endrun, a
CAM-specific routine, to instead pass the string

     "See atm.log for error description"

to the shr_sys_abort routine whne endrun is called with no parameters.

Fixes #2498

[BFB]