COMPOSE: Add nearest-point calculation.
If semi_lagrange_nearest_point_lev is set to >= 1, then in any lev <= this
parameter, the SL advection method is permitted to use an (approximate) nearest
point to the target cell's halo if the departure point is outside of the halo.

Fix a bug in the get_src_cell trial > 1 backup calculation.

Fix output bugs in the get_src_cell failure message.

Write unit tests for the nearest-point calculations.

Update prim_main to hand off to the SL convergence tester for all available
convergence-tested SL algorithms.

Fix MPI type in the SL-related control-mod values. This bug hasn't affected us,
but the type was clearly wrong, logical rather than integer.