Migrate atmospheric forcing height to topounit data structure
Atmospheric forcing height from coupler is used to fill the zbot variable of the
topounit atmospheric state data structure. Although there are references later in
the code to separate forcing heights for temperature, humidity, and winds, there
is only one quantity passed from the coupler. For that reason, only one quantity
is defined in the topounit data structure.
Testing was performed with the full e3sm_land_developer suite.

[BFB] - Bit-For-Bit
L2-178