Extract common portions of regex
The main purpose of this is: assert_phase_absent could wrongly pass if
the format of the TestStatus output changed without that method's regex
changing. By making its regex shared as much as possible with the regex
in assert_status_of_phase, we decrease the chances of these false
passes.