pylint fixes
assertNotRegexpMatches is deprecated in python3.5 and later, leading to
a pylint warning. six doesn't provide an alias for
assertNotRegexpMatches / assertNotRegex, so I have written our own,
following the style used by six. I kept this in a separate module to
make it easier to update to later versions of six in the future.