Revert to old radiation interface
Revert to old radiation call interface. This is done to minimize interference
with master and RRTMG when submitting the PR for these changes, and because
I was having a hard time debugging run errors on cori-knl with the changes
I had made. This reverts to using comsrf for some fields, passing the net
fluxes in and out of radiation_tend rather than putting them on the pbuf,
and reverting to the clunky way of passing the is_cmip6_volc flag around. Builds
with RRTMG now run on cori-knl again though, and RRTMGP builds but fails at the
read stage for the coefficients files for some reason.