Do not report BFAIL for a test that is missing a file
If a test had both (a) a missing file in the baseline that was present
in the test, and (b) a missing file in the test that was present in the
baseline, this was previously reported as a BFAIL rather than a true
failure. This commit fixes this problem, reporting it as a true failure.