Revert "Automatic code formatting commit"
This reverts commit a196719f732cd82a7208f3469f7d9b78eed4d649.

Revert an accidental automated format commit and push to master.

We should discuss this part of the dev-ops setup before using it. In addition,
there are at least two formatting conventions currently used that I believe are
highly nonstandard: no space between a control keyword and open parens, and
putting the pointer annotation on the variable rather than the type.