Add /cluster/torque/bin to path on hobart
This is needed in order to submit jobs from the compute nodes, via
qsub. (I want to be able to do this so that we can build and submit
create_test jobs on the compute nodes.) Somehow we've been able to do
resubmits on hobart, but I can't figure out how that is working; Mark
Moore suggested simply adding this to our path.