HOMME: Fix array sizes to mirror packs.
ptens from iptens was a bug when VECTOR_SIZE doesn't divide
NUM_PHYSICAL_LEV. The other two changes are not strictly necessary, but it's
good to align each column according to the vector size.

Modify the limiters_ut test to have 13 physical levels. This tests pack
arithmetic correctness. With this mod, the test failed before the fix above; now
it passes.