mass balance bug fix for betr c or c-n only run
This fixed some numerical bug that led to mass balance error
for nitrogen and phosphorus. The cause was due to the
limited precision that a computer's memory can hold. The
strategy is to apply the fluxes one by one rather than
by first taking their summary and then applying all at once.
No change is expected to default elm.